+++ crates/test/src/error.rs
@@ -1,7 +1,6 @@
 use std::io;
 
 use smplx_sdk::provider::ProviderError;
-use smplx_sdk::signer::SignerError;
 
 use smplx_regtest::error::RegtestError;
 
@@ -13,12 +12,12 @@ pub enum TestError {
     #[error(transparent)]
     Provider(#[from] ProviderError),
 
-    #[error(transparent)]
-    Signer(#[from] SignerError),
-
     #[error("Failed to deserialize config: '{0}'")]
     ConfigDeserialize(#[from] toml::de::Error),
 
     #[error("io error occurred: '{0}'")]
     Io(#[from] io::Error),
+
+    #[error("Network name should either be `Liquid`, `LiquidTestnet` or `ElementsRegtest`, got: {0}")]
+    BadNetworkName(String),
 }